UX – move resolve duplicate warning button + fix copy
