+++ packages/tokens-studio-for-figma/src/i18n/lang/en/tokens.json
@@ -128,8 +128,9 @@
     }
   },
   "resolveDuplicateTokensModal": {
-    "title": "Resolve Duplicate Groups",
-    "description": "Duplicate tokens (with the same token name) were found, please choose which token to keep. There should only be one token resolved for each token name."
+    "title": "Resolve Duplicate Tokens",
+    "description": "Duplicate tokens (with the same token name) were found, please choose which token to keep. There should only be one token resolved for each token name.",
+    "submit": "Resolve Duplicates"
   },
   "emptyGroups": "empty groups",
   "duplicateGroup": "Duplicate Group",
